fix: rename files in component subfolders
